Address PR review: split postprocess/display, move imports to top-level, align config placement, update README format
