--- local_resources/yolov4_license_plate.json
@@ -1,16 +1,0 @@
-{
-  "iou_threshold": 0.45,
-  "detection_threshold": 0.3,
-  "output_activation": "sigmoid",
-  "label_offset": 1,
-  "max_boxes": 200,
-  "anchors": [
-    [81, 82, 135, 169],
-    [344, 319, 10, 14],
-    [23, 27, 37, 58]
-  ],
-  "labels": [
-    "unlabeled",
-    "license_plate"
-  ]
-}
